Extracted atom stuff from Db funcs
