+++ mobile-app/src/firebase/api.ts
@@ -193,7 +193,7 @@ export function subscribeToFriendList(setter: SetFunction): Unsubscribe {
 
   const currentUserId = atomStore.get(currentUserIdAtom);
   fetchFriendData({ userId: currentUserId }).then(setter);
-  return () => {};
+  return () => { };
 }
 
 export async function fetchCommonHabitIds({
@@ -417,58 +417,37 @@ export async function searchUsersInDb({
 }
 
 export async function updateProfileDataInDB(
-  profileFormDataAtom: PrimitiveAtom<{
-    displayName: string;
-    username: string;
-  }>,
-  currentUserAtom: PrimitiveAtom<userWithIdT>,
+  currentUserID: string,
+  newDataForDb: userT
 ) {
-  async () => {
-    const profileFormData = useAtomValue(profileFormDataAtom);
-    const [userData, setUserData] = useAtom(currentUserAtom);
-
+  {
     // if (auth.currentUser == null) {
     //   console.log("Cannot edit profile: User not logged in")
     //   return;
     // }
 
-    //Ensure user is unique
-    const usersRef = collection(firestore, "users");
-    const q = query(
-      usersRef,
-      where("username", "==", profileFormData.username),
-    );
-    const querySnapshot = await getDocs(q);
-
-    // Check if were changing username, if so check if username already exist in DB
-    if (
-      userData.username !== profileFormData.username &&
-      !querySnapshot.empty
-    ) {
-      console.log("Error: Username already taken"); // Display this on screen instead
-      profileFormData.username = "";
-      return;
-    }
-
-    // On success
-    const newDataForAtom: userWithIdT = {
-      ...userData,
-      ...profileFormData,
-    };
-
-    const newDataForDb: userT = {
-      username: profileFormData.username,
-      displayName: profileFormData.displayName,
-      picture: userData.picture,
-      createdAt: userData.createdAt,
-    };
-
     //Push Data to DB
-    const userDocRef = doc(firestore, "users", userData.id); //change to auth.currentUser.uid when auth is fixed
+    const userDocRef = doc(firestore, "users", currentUserID); //change to auth.currentUser.uid when auth is fixed
     await setDoc(userDocRef, newDataForDb);
-
-    //Update the current atoms accordingly
-    setUserData(newDataForAtom);
-    router.back();
   };
+}
+
+export async function newUsernameIsUnique(existing_username: string, new_username: string): Promise<boolean> {
+  //Ensure user is unique
+  const usersRef = collection(firestore, "users");
+  const q = query(
+    usersRef,
+    where("username", "==", new_username),
+  );
+  const querySnapshot = await getDocs(q);
+
+  // Check if were changing username, if so check if username already exist in DB
+  if (
+    new_username !== existing_username &&
+    !querySnapshot.empty
+  ) {
+    console.log("Error: Username already taken"); // Display this on screen instead
+    return false;
+  }
+  return true;
 }
